requirements.txt
accelerate==0.27.0
datasets==2.15.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
faiss-cpu==1.8.0.post1
nltk==3.8.1
openai==1.54.3
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.12.0
selfcheckgpt==0.1.4
sentence-transformers==2.3.0
sentencepiece==0.2.0
tiktoken==0.7.0
tokenizers==0.20.3
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.1
transformers==4.46.2
tiktoken
python-dotenv==1.0.1
openpyxl==3.1.2
pandas==2.2.2